share model cache helpers
